Clean up duplicate types
